Split out settings container
